+++ 2DShooterProject/2DShooter/Game.cpp
@@ -24,7 +24,7 @@ Game::Game() :
 	m_bRunning(false),
 	m_pGameStateMachine(0),
 	m_playerLives(3),
-	m_scrollSpeed(0.8),
+	m_scrollSpeed((float)0.8),
 	m_bLevelComplete(false),
 	m_bChangingState(false)
 {
