+++ 2DShooterProject/2DShooter/ScrollingBackground.cpp
@@ -9,17 +9,17 @@ void ScrollingBackground::load(std::unique_ptr<LoaderParams> const& pParams)
     m_scrollSpeed = 1;
 
     m_srcRect1.x = 0;
-    m_destRect1.x = m_position.getX();
+    m_destRect1.x = (int)m_position.getX();
     m_srcRect1.y = 0;
-    m_destRect1.y = m_position.getY();
+    m_destRect1.y = (int)m_position.getY();
 
     m_srcRect1.w = m_destRect1.w = m_srcRect2Width = m_destRect1Width = m_width;
     m_srcRect1.h = m_destRect1.h = m_height;
 
     m_srcRect2.x = 0;
-    m_destRect2.x = m_position.getX() + m_width;
+    m_destRect2.x = (int)m_position.getX() + m_width;
     m_srcRect2.y = 0;
-    m_destRect2.y = m_position.getY();
+    m_destRect2.y = (int)m_position.getY();
 
     m_srcRect2.w = m_destRect2.w = m_srcRect2Width = m_destRect2Width = 0;
     m_srcRect2.h = m_destRect2.h = m_height;
@@ -53,17 +53,17 @@ void ScrollingBackground::update()
         if (m_destRect2.w >= m_width)
         {
             m_srcRect1.x = 0;
-            m_destRect1.x = m_position.getX();
+            m_destRect1.x = (int)m_position.getX();
             m_srcRect1.y = 0;
-            m_destRect1.y = m_position.getY();
+            m_destRect1.y = (int)m_position.getY();
 
             m_srcRect1.w = m_destRect1.w = m_srcRect2Width = m_destRect1Width = m_width;
             m_srcRect1.h = m_destRect1.h = m_height;
 
             m_srcRect2.x = 0;
-            m_destRect2.x = m_position.getX() + m_width;
+            m_destRect2.x = (int)m_position.getX() + m_width;
             m_srcRect2.y = 0;
-            m_destRect2.y = m_position.getY();
+            m_destRect2.y = (int)m_position.getY();
 
             m_srcRect2.w = m_destRect2.w = m_srcRect2Width = m_destRect2Width = 0;
             m_srcRect2.h = m_destRect2.h = m_height;
